refactor : Participant 상속을 통한 Dealer-Player 간 중복 제거
