+++ src/main/java/model/Dealer.java
@@ -1,34 +1,13 @@
 package model;
 
-import java.util.List;
 
-public class Dealer {
-    private final ParticipantHand participantHand;
-
+public class Dealer extends Participant{
     public Dealer() {
-        this.participantHand = new ParticipantHand();
-    }
-
-    public void receiveCard(Card card) {
-        participantHand.add(card);
+        super();
     }
 
     public boolean checkScoreUnderSixteen() {
-        return participantHand.checkScoreBelow(16);
-    }
-
-    public void dealInitialCards(Deck deck) {
-        for (int i = 0; i < 2; i++) {
-            receiveCard(deck.pick());
-        }
-    }
-
-    public List<Card> getHandCards() {
-        return participantHand.getCards();
-    }
-
-    public ParticipantHand getParticipantHand() {
-        return participantHand;
+        return getParticipantHand().checkScoreBelow(16);
     }
 }
