+++ src/test/java/model/ParticipantHandTest.java
@@ -72,31 +72,31 @@ class ParticipantHandTest {
         Assertions.assertTrue(burst);
     }
 
-    @Test
-    @DisplayName("ACE가 한장일때 ACE를 뺸 나머지의 합이 10 이하일경우 true")
-    void ACE를_제외한_나머지의_합이_10_이하이다(){
-        ParticipantHand participantHand = new ParticipantHand();
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.EIGHT, CardSuit.HEART));
-        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
+//    @Test
+//    @DisplayName("ACE가 한장일때 ACE를 뺸 나머지의 합이 10 이하일경우 true")
+//    void ACE를_제외한_나머지의_합이_10_이하이다(){
+//        ParticipantHand participantHand = new ParticipantHand();
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.EIGHT, CardSuit.HEART));
+//        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
+//
+//        boolean result = participantHand.isAceElevenPossible();
+//
+//        assertTrue(result);
+//    }
 
-        boolean result = participantHand.isAceElevenPossible();
-
-        assertTrue(result);
-    }
-
-    @Test
-    @DisplayName("ACE가 한장일 때 ACE를 뺸 나머지의 합이 11 이샹일경우 false")
-    void ACE를_제외한_나머지의_합이_11_이상이다(){
-        ParticipantHand participantHand = new ParticipantHand();
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.NINE, CardSuit.HEART));
-        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
-
-        boolean result = participantHand.isAceElevenPossible();
-
-        assertFalse(result);
-    }
+//    @Test
+//    @DisplayName("ACE가 한장일 때 ACE를 뺸 나머지의 합이 11 이샹일경우 false")
+//    void ACE를_제외한_나머지의_합이_11_이상이다(){
+//        ParticipantHand participantHand = new ParticipantHand();
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.NINE, CardSuit.HEART));
+//        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
+//
+//        boolean result = participantHand.isAceElevenPossible();
+//
+//        assertFalse(result);
+//    }
 
     @Test
     @DisplayName("ACE를 뺀 나머지의 합이 10 이하일 경우 ACE 한 장을 11로 계산한다")
@@ -111,35 +111,37 @@ class ParticipantHandTest {
 
         assertEquals(expect, result);
     }
-    @Test
-    @DisplayName("ACE_한_장을_ACE의_최대점수인_11점으로 바뀔 수 있는지 확인한다")
-    void ACE_한_장을_ACE의_최대점수인_11점으로_바뀔_수_있는지_확인한다(){
-        ParticipantHand participantHand = new ParticipantHand();
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.SIX, CardSuit.HEART));
-        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
 
-        boolean result = participantHand.isAceElevenPossible();
+//    @Test
+//    @DisplayName("ACE_한_장을_ACE의_최대점수인_11점으로 바뀔 수 있는지 확인한다")
+//    void ACE_한_장을_ACE의_최대점수인_11점으로_바뀔_수_있는지_확인한다(){
+//        ParticipantHand participantHand = new ParticipantHand();
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.SIX, CardSuit.HEART));
+//        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
+//
+//        boolean result = participantHand.isAceElevenPossible();
+//
+//        assertTrue(result);
+//    }
 
-        assertTrue(result);
-    }
+//    @Test
+//    @DisplayName("ACE_한_장을_ACE의_최대점수인_11점으로 바뀔 수 없는 경우 확인")
+//    void ACE_한_장을_ACE의_최대점수인_11점으로_바뀔_수_없는_경우_확인(){
+//        ParticipantHand participantHand = new ParticipantHand();
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
+//        participantHand.add(new Card(CardRank.SEVEN, CardSuit.HEART));
+//        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
+//
+//        boolean result = participantHand.isAceElevenPossible();
+//
+//        assertFalse(result);
+//    }
 
-    @Test
-    @DisplayName("ACE_한_장을_ACE의_최대점수인_11점으로 바뀔 수 없는 경우 확인")
-    void ACE_한_장을_ACE의_최대점수인_11점으로_바뀔_수_없는_경우_확인(){
-        ParticipantHand participantHand = new ParticipantHand();
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.ACE, CardSuit.DIAMOND));
-        participantHand.add(new Card(CardRank.SEVEN, CardSuit.HEART));
-        participantHand.add(new Card(CardRank.TWO, CardSuit.DIAMOND));
-
-        boolean result = participantHand.isAceElevenPossible();
-
-        assertFalse(result);
-    }
     /**
      * 나머지의 합이 10 이하다
      *  = a의 값이 11 이상일 수 있다.
